Clears LC_ALL during database creation [#68384446]
This allows the template databases to be created with the UTF-8/Unicode
encoding.  The 'command' and other resources use MixLib::ShellOut to execute
the commands on the shell, but for some legacy ruby reason they default to
setting the 'LC_ALL' env variable to 'C' during execution.  setting it to nil
essentially overrides that behaviour allowing the users environment to take
precedence.

There is an open bug about this issue:
https://tickets.opscode.com/browse/CHEF-2784